Wait for post migrate state after migration
During migration QEMU reports it is in the paused state. At the end of the
migration it transitions to the postmigrate state. It appears that there is a
race between query-migrate reporting that the migration is finished and QEMU
transitioning to the postmigrate state. Therefor it is possible to put QEMU
into the running state (with cont) before it tries to transition to
postmigrate. So when it then tries to transition to postmigrate it reports
that this is an invalid state transition.

Therefor we wait for QEMU to transition from the paused state before issuing
cont.